fix(mailbox): use fxa-friendly validation regexes
In fxa-email-service, we validate email addresses before they reach this
crate. Our validation layer permits UTF-8 alphanumerics from other
character sets in the local part of the address.

Older versions of this crate parsed those addresses okay but as of
version 2.2 they are rejected. This commit changes it to use the same
validation regexes as fxa-email-service.